+++ test/linkedlist.c
@@ -3,7 +3,9 @@
 
 #include <CUTE/cute.h>
 #include <clog.h> /* for logging macros */
+#include <stddef.h> /* for size_t */
 #include <stdlib.h> /* for NULL */
+#include <unistd.h> /* for ssize_t */
 
 
 
@@ -15,7 +17,7 @@ CUTE_TestCase *case_linkedlist;
 static LinkedList *llist;
 
 static int VALUES[] = {42, 3, 7, 13, 6};
-static const unsigned int INT_LINKED_LIST_SIZE = 5;
+static const size_t INT_LINKED_LIST_SIZE = 5;
 
 
 extern _Bool equal_as_ints(const data_t*, const data_t*);
@@ -28,7 +30,7 @@ static void init(void) {
 	verbose("llist = ll_new()");
 	llist = ll_new();
 	CUTE_assertNotEquals(llist, NULL);
-	for(unsigned int i = 0; i < INT_LINKED_LIST_SIZE; ++i) {
+	for(size_t i = 0; i < INT_LINKED_LIST_SIZE; ++i) {
 		ll_add(llist, i, &VALUES[i]);
 	}
 }
@@ -41,14 +43,14 @@ static void cleanup(void) {
 
 static void test_ll_len__empty(void) {
 	LinkedList *empty_llist;
-	unsigned int got;
+	size_t got;
 	notice("test ll_len -- empty list");
 	info("empty_llist = ll_new()");
 	empty_llist = ll_new();
 	verbose("ll_len(empty_llist)");
 	info("expected: 0");
 	got = ll_len(empty_llist);
-	info("got     : %u", got);
+	info("got     : %zu", got);
 	CUTE_assertEquals(got, 0);
 	ll_free(empty_llist);
 	verbose("OK");
@@ -61,7 +63,7 @@ static void test_ll_append(void) {
 	notice("tests ll_append");
 	verbose("empty_llist = ll_new()");
 	empty_llist = ll_new();
-	for(unsigned int n = 0; n < INT_LINKED_LIST_SIZE; ++n) {
+	for(size_t n = 0; n < INT_LINKED_LIST_SIZE; ++n) {
 		param = VALUES + n;
 		verbose("ll_append(empty_llist, %p)", param);
 		expected = n;
@@ -76,12 +78,12 @@ static void test_ll_append(void) {
 }
 
 static void test_ll_len__full(void) {
-	unsigned int got;
+	size_t got;
 	notice("test ll_len -- list full");
 	verbose("ll_len(llist)");
-	info("expected: %u", INT_LINKED_LIST_SIZE);
+	info("expected: %zu", INT_LINKED_LIST_SIZE);
 	got = ll_len(llist);
-	info("got     : %u", got);
+	info("got     : %zu", got);
 	CUTE_assertEquals(got, INT_LINKED_LIST_SIZE);
 	verbose("OK");
 }
@@ -89,8 +91,8 @@ static void test_ll_len__full(void) {
 static void test_ll_get__valid(void) {
 	data_t *expected, *got;
 	notice("test ll_get -- valid indices");
-	for(unsigned int index = 0; index < INT_LINKED_LIST_SIZE; ++index) {
-		verbose("ll_get(llist, %u)", index);
+	for(size_t index = 0; index < INT_LINKED_LIST_SIZE; ++index) {
+		verbose("ll_get(llist, %zu)", index);
 		expected = VALUES + index;
 		info("expected: %p", expected);
 		got = ll_get(llist, index);
@@ -102,17 +104,17 @@ static void test_ll_get__valid(void) {
 }
 
 static void test_ll_get__invalid(void) {
-	const unsigned int invalid_indices[3] = {
+	const size_t invalid_indices[3] = {
 		ll_len(llist),
 		ll_len(llist) + 1,
 		42
 	};
 	data_t *got;
-	unsigned int index;
+	size_t index;
 	notice("test ll_get -- invalid indices");
-	for(unsigned int i = 0; i < 3; ++i) {
+	for(size_t i = 0; i < 3; ++i) {
 		index = invalid_indices[i];
-		verbose("ll_get(llist, %u)", index);
+		verbose("ll_get(llist, %zu)", index);
 		info("expected: (nil)");
 		got = ll_get(llist, index);
 		info("got     : %p", got);
@@ -125,9 +127,9 @@ static void test_ll_get__invalid(void) {
 static void test_ll_set__valid(void) {
 	static int extra = 64;
 	data_t *const param = &extra;
-	const unsigned int index = 4;
+	const size_t index = 4;
 	notice("test ll_set -- valid index");
-	verbose("ll_set(llist, %u, %p)", index, param);
+	verbose("ll_set(llist, %zu, %p)", index, param);
 	info("expected errno: 0");
 	ll_set(llist, index, param);
 	info("actual errno  : %d", errno);
@@ -138,16 +140,16 @@ static void test_ll_set__valid(void) {
 static void test_ll_set__invalid(void) {
 	static int extra = 23;
 	data_t *const param = &extra;
-	const unsigned int invalid_indices[3] = {
+	const size_t invalid_indices[3] = {
 		ll_len(llist),
 		ll_len(llist) + 1,
 		73 /* big "arbitrary" value */
 	};
-	unsigned int index;
+	size_t index;
 	notice("tests ll_set -- invalid indices");
-	for(unsigned int i = 0; i < 3; ++i) {
+	for(size_t i = 0; i < 3; ++i) {
 		index = invalid_indices[i];
-		verbose("ll_set(llist, %u, %p)", index, param);
+		verbose("ll_set(llist, %zu, %p)", index, param);
 		info("expected errno: %d", ERANGE);
 		ll_set(llist, index, param);
 		info("actual errno  : %d", errno);
@@ -160,9 +162,9 @@ static void test_ll_add__valid(void) {
 	static int extra = 8;
 	data_t *const param = &extra;
 	int expected, got;
-	const unsigned int index = 3;
+	const size_t index = 3;
 	notice("test ll_add -- valid index");
-	info("ll_add(llist, %u, %p)", index, param);
+	info("ll_add(llist, %zu, %p)", index, param);
 	expected = (signed)index;
 	info("expected: %d", expected);
 	got = ll_add(llist, index, param);
@@ -175,20 +177,20 @@ static void test_ll_add__valid(void) {
 static void test_ll_add__invalid(void) {
 	int extra = 9;
 	data_t *const param = &extra;
-	const unsigned int invalid_indices[3] = {
+	const size_t invalid_indices[3] = {
 		ll_len(llist) + 1,
 		ll_len(llist) + 2,
 		668
 	};
-	int got;
-	unsigned int index;
+	ssize_t got;
+	size_t index;
 	notice("test ll_add -- invalid indices");
-	for(unsigned int i = 0; i < 3; ++i) {
+	for(size_t i = 0; i < 3; ++i) {
 		index = invalid_indices[i];
-		info("ll_add(llist, %u, %p)", index, param);
+		info("ll_add(llist, %zu, %p)", index, param);
 		info("expected: -1");
 		got = ll_add(llist, index, param);
-		info("got     : %d", got);
+		info("got     : %zd", got);
 		CUTE_assertEquals(got, -1);
 		CUTE_assertErrnoEquals(ERANGE);
 	}
@@ -197,9 +199,9 @@ static void test_ll_add__invalid(void) {
 
 static void test_ll_drop__valid(void) {
 	data_t *expected, *got;
-	const unsigned int index = 0;
+	const size_t index = 0;
 	notice("test ll_drop -- valid index");
-	info("ll_drop(llist, %u)", index);
+	info("ll_drop(llist, %zu)", index);
 	expected = VALUES + index;
 	info("expected: %p", expected);
 	got = ll_drop(llist, index);
@@ -210,17 +212,17 @@ static void test_ll_drop__valid(void) {
 }
 
 static void test_ll_drop__invalid(void) {
-	const unsigned int invalid_indices[3] = {
+	const size_t invalid_indices[3] = {
 		ll_len(llist),
 		ll_len(llist) + 1,
 		99
 	};
 	data_t *got;
-	unsigned int index;
+	size_t index;
 	notice("test ll_drop -- invalid indices");
-	for(unsigned int i = 0; i < 3; ++i) {
+	for(size_t i = 0; i < 3; ++i) {
 		index = invalid_indices[i];
-		info("ll_drop(llist, %u)", index);
+		info("ll_drop(llist, %zu)", index);
 		info("expected: (nil)");
 		got = ll_drop(llist, index);
 		info("got     : %p", got);
@@ -233,10 +235,10 @@ static void test_ll_drop__invalid(void) {
 static void test_ll_swap__valid(void) {
 	static int extra = -32;
 	data_t *const param = &extra;
-	const unsigned int index = 2;
+	const size_t index = 2;
 	data_t *expected, *got;
 	notice("test ll_swap -- valid index");
-	verbose("ll_swap(llist, %u, %p)", index, param);
+	verbose("ll_swap(llist, %zu, %p)", index, param);
 	expected = VALUES + index;
 	info("expected: %p = %d", expected, *(int*)expected);
 	got = ll_swap(llist, index, param);
@@ -249,17 +251,17 @@ static void test_ll_swap__valid(void) {
 static void test_ll_swap__invalid(void) {
 	int extra = 4;
 	data_t *const param = &extra;
-	const unsigned int invalid_indices[3] = {
+	const size_t invalid_indices[3] = {
 		ll_len(llist),
 		ll_len(llist) + 1,
 		13
 	};
 	data_t *got;
-	unsigned int index;
+	size_t index;
 	notice("test ll_swap -- invalid indices");
-	for(unsigned int i = 0; i < 3; ++i) {
+	for(size_t i = 0; i < 3; ++i) {
 		index = invalid_indices[i];
-		verbose("ll_swap(llist, %u, %p)", index, param);
+		verbose("ll_swap(llist, %zu, %p)", index, param);
 		info("expected: (nil)");
 		got = ll_swap(llist, index, param);
 		info("got     : %p", got);
